fix: Ensures the correct SCM domain is found for uploading zip package (#178)
The SCM domain is different based on the type of function app deployment. In a typical function app, the SCM domain is always {appName}.scm.azurewebsites.net.

When deploying a function into an App Service Environment (ASE) the SCM domain is a sub domain off of the ASE, ex) {appName}.scm.{asePath}.p.azurewebsites.net

This fix addresses the difference in finding the correct SCM domain to deploy the zip package.