refactor: Moved hooks requiring login to separate file (#206)
Rather than changing the `AzureLoginPlugin` to add `"before:new:hook": this.login.bind(this)` every time a new hook is added that needs authentication, this allows for the hooks (without `before:`) to be placed in a separate file. These are loaded in the constructor of the plugin, which sets up the proper hooks.