fix: Default operation names no longer cause collisions (#332)
APIM uses the operation name as the key when upserting operations into an API. By default we were using the function name as the operation name. During default configuration inference with multiple HTTP methods this caused a collision resulting in the last operation in always one.

This fix generates a unique operation name taking into account the function name + the HTTP method.